major refactoring of spatial
