+++ src/test/scala/org/zouzias/spark/lucenerdd/spatial/point/PointLuceneRDDSpec.scala
@@ -20,6 +20,7 @@ import com.holdenkarau.spark.testing.SharedSparkContext
 import org.scalatest.{BeforeAndAfterEach, FlatSpec, Matchers}
 import org.zouzias.spark.lucenerdd.implicits.LuceneRDDImplicits._
 import org.zouzias.spark.lucenerdd.models.SparkDoc
+import org.zouzias.spark.lucenerdd.spatial.core.SpatialLuceneRDD
 import org.zouzias.spark.lucenerdd.spatial.implicits.PointLuceneRDDImplicits._
 
 class PointLuceneRDDSpec extends FlatSpec
@@ -34,7 +35,7 @@ class PointLuceneRDDSpec extends FlatSpec
   val Toronto = ((-79.4, 43.7), "Toronto")
   val k = 5
 
-  var pointLuceneRDD: PointLuceneRDD[_, _] = _
+  var pointLuceneRDD: SpatialLuceneRDD[_, _] = _
 
   override def afterEach() {
     pointLuceneRDD.close()
